move mp4 methods out of demuxer
